EOD 12/9/2020; Reading Maps From Files
Maps can be read in from json files. A lot of rearrangement of files.

Still very much a work in progress. In this state, the game is unplayable, because the map loaded is too tiny to contain the units that will spawn on it. The game does not fall back on a working state, it merely does not function.